requirements.txt
asgiref==3.8.1
attrs==25.3.0
diff-match-patch==20241021
Django==5.2.3
django-constance==4.3.2
django-debug-toolbar==5.2.0
django-extensions==4.1
django-import-export==4.3.8
djangorestframework==3.16.0
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
python-decouple==3.8
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.1
sqlparse==0.5.3
tablib==3.8.0
typing_extensions==4.14.0
uritemplate==4.2.0
